+++ src/nfa.c
@@ -5,7 +5,7 @@
 
 void free_nfa(struct nfa* nfa) {
 	list_clear(&nfa->exits);
-	int i, j;
+	uintptr_t i, j;
 	for (i = 0; i < nfa->node_count; ++i) {
 		struct nfa_node* node = nfa->nodes[i];
 		for (j = 0; j < node->edge_count; ++j) {
